Shuffle code around into separate REPL package.
