+++ src/main/kotlin/com/jetbrains/micropython/ui/MicroPythonToolWindowFactory.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.content.ContentFactory
+import com.jetbrains.micropython.repl.ToolWindowReplTab
 import com.jetbrains.micropython.settings.firstMicroPythonFacet
 
 class MicroPythonToolWindowFactory : ToolWindowFactory, DumbAware {
