+++ src/main/resources/META-INF/plugin.xml
@@ -76,8 +76,8 @@
                          id="com.jetbrains.micropython.configurable"
                          provider="com.jetbrains.micropython.settings.MicroPythonConfigurableProvider"/>
     <projectService serviceImplementation="com.jetbrains.micropython.settings.MicroPythonDevicesConfiguration"/>
-    <projectService serviceImplementation="com.jetbrains.micropython.run.DeviceCommsManager"/>
-    <stepsBeforeRunProvider implementation="com.jetbrains.micropython.run.StopReplBeforeRunTaskProvider"/>
+    <projectService serviceImplementation="com.jetbrains.micropython.repl.DeviceCommsManager"/>
+    <stepsBeforeRunProvider implementation="com.jetbrains.micropython.repl.StopReplBeforeRunTaskProvider"/>
     <localInspection language="Python"
                      displayName="MicroPython requirements inspection"
                      groupName="MicroPython"
